Initial commit.
I have basically copied the code for their helloworld, smith-waterman and k-means
clustering algorithms. Obviously I'm recompiling it and putting it in my own jar
file. For the helloworld and smithwaterman, I did not change anything. However, for
the k-means algorithm, I have changed a few things.

Basically, I am writing out txt files (something which I do not think was done
before). The files show the input data (which is generated within the running
program), the values of each of the centroids after each iteration and the final
cluster membership (i.e. which vector is in which cluster).

I can potentially claim this as a useful addition in that k-means clustering is now
usable by a Ciel user -- i.e. it outputs the final membership for each vector.