Generate random data for canopy clustering.
Started programming canopy clustering. There is a CanopyInitTask which
basically is the controller. If the user does not pass in a file path pointing
to a data file, then it invokes the CanopyDataGenerator task which creates the
data and returns an object. At the moment, I'm blocking until that object has
finished being made, but the long term plan is to pass a reference to mappers.