Wrote the Reducer class.
Completed canopy clustering. The reducer takes all of the output centres
from the different mappers and re-runs canopy clustering. The ouput is written
as a WritableReference and so would only be avialable to further Ciel tasks.
At the moment, I am writing results out to a text file as well for validation
purposes, but this must be stopped for benchmarking.