Generalized the DataGenerator.
Since a new algorithm in a new package is required, but it still needs to use
the same data generation, it was decided to extract the data generation from
the canopy package and put it in its own package. This caused the rename. It
was also required to edit the CanopyInitTask to work with the new package.