A class which makes an initial list of clusters.
At the moment, the class simply makes a new cluster for each data point. So,
this means that the centre of each of the original clusters is the original
data point. While I am not sure that this is the correct way to initialize the
clusters, it does not seem like such a bad idea because manay of these initial
clusters will surely merge over the iterations.