Added MSReducer and completed mean shift clustering.
Wrote the MSReducer class. Based on whether there is convergance or not, it
can generate a series of map tasks followed by a further reducer. In other
words, it controls iterations or allows for it to work iteratively.
MSInitTask was also edited to work with the new MSReducer.
The mean shift clustering algorithm is now (theoretically) complete.