Apply small changes to datageneration, meanshift and canopy code.
Edited a number of classes so that they no longer create text files. This was
primarily used for debugging, but it required one to create a data folder in
/tmp. This was unsatisfactory.

Removed the seeds in the random number generation. Again, this was
primarily used for testing. Now, random number are generated using the current
time as a seed.